Add ability to create tasks
This is mostly to serve as a "reminder" use-case - when you
think of something and want to jot it down somewhere you won't
lose track of it.

As a result, it's pretty basic: you only input a task name.
It sets the due date to today so you notice it next time you
open Asana, and can triage it to the right place.